requirements.txt
pyaoscx
PyYAML
requests
tftpy
urllib3
Jinja2
pycentral